pyproject.toml
[tool.poetry]
name = "voice-agent"
version = "0.1.0"
description = ""
authors = ["Alan Bohannon <alanbohannon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pipecat = "^0.3.0"
dotenv = "^0.9.9"
loguru = "^0.7.3"
pipecat-ai-cli = "^0.1.6"
pipecat-ai = {extras = ["webrtc","daily","silero","deepgram","openai","cartesia","local-smart-turn-v3","runner","local-smart-turn-v3","elevenlabs"], version = "^0.0.91"}
websockets = "^15.0.1"
nylas = "^6.13.1"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
